Skip prefix-escaping symlinks and tighten glob capability tests
An escaping symbolic link aborted the whole expansion: `Dir::metadata`
follows links inside the sandbox, so a match such as `src/a.c` pointing
at `../vendor/a.c` raised `PermissionDenied` ("a path led outside of the
filesystem") and failed `glob_paths`. Such an entry names no file the
capability can reach, so skip it and keep the traversal going, leaving
genuine metadata errors on non-symlinks fatal.

`glob_paths_returns_empty_when_literal_prefix_missing` passed against
the previous ambient implementation too, and no case covered a prefix
naming a regular file. Replace it with a direct assertion that
`open_root_dir` declines both, and add one exercising the relativisation
`GlobRoot::metadata` performs: a match under the prefix resolves, while
a sibling outside it is rejected as `InvalidInput`.

Share the duplicated temporary tree through a `scoped_tree` rstest
fixture, and split the test module into `pattern`, `expansion`, and
`capability` submodules; the additions took it past the 400-line module
limit.

Document the user-visible behaviour in the users' guide: parent-relative
patterns, the empty result for a missing or non-directory literal
prefix, the skipped escaping symlink, and the capability scope itself
under the safety boundary.
