+++ CHANGELOG.md
@@ -64,7 +64,17 @@
 - Open the capability used for glob metadata checks at the pattern's longest
   literal directory prefix rather than at the filesystem root or the working
   directory, so glob expansion holds only the authority its pattern can reach;
-  a pattern whose literal prefix is missing now expands to no matches
+  a pattern whose literal prefix is missing or names something other than a
+  directory now expands to no matches, and a match that is a symbolic link
+  resolving outside that prefix is skipped rather than failing the expansion
+  ([#173](https://github.com/leynos/netsuke/issues/173))
+
+
+### Fixed
+
+- Expand parent-relative glob patterns such as `glob('../shared/*.h')`; their
+  matches previously reached the working-directory capability as `../…` and
+  were rejected as sandbox escapes
   ([#173](https://github.com/leynos/netsuke/issues/173))
 
 ### Removed
